+++ test/test_cli_union.sh
@@ -92,7 +92,7 @@ validate("Validate changes"), cli_validate();
 commit("Commit the changes"), cli_commit();
 quit("Quit"), cli_quit();
 show("Show a particular state of the system"){
-    auto("Show expand") @datamodel, cli_show_auto("running", "xml", "set");  
+    auto("Show expand") @datamodel, cli_show_auto("running", "xml");  
     candidate("Show configuration"), cli_auto_show("datamodel", "candidate", "xml", false, false);
     running("Show configuration"), cli_auto_show("datamodel", "running", "xml", false, false);
 }